move p2p to trace.py
